+++ transactions/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 
 class Transaction(models.Model):
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.IntegerField()
     checkout_id = models.CharField(max_length=100, unique=True)
     mpesa_code = models.CharField(max_length=20, unique=True)
     phone_number = models.CharField(max_length=15)
